+++ assets/utilities/component/custom-prop-types.js
@@ -4,10 +4,12 @@ const buttonStyleTypes = React.PropTypes.oneOf(['success', 'info', 'warning', 'd
 const buttonAligns = React.PropTypes.oneOf(['left', 'right']);
 const gridRowAlignments = React.PropTypes.oneOf(['left', 'center', 'right', 'spaced', 'justify']);
 const cardStyleTypes = React.PropTypes.oneOf(['success', 'info', 'warning', 'danger']);
+const progressBarStyleTypes = React.PropTypes.oneOf(['success', 'info', 'warning', 'danger']);
 
 export default {
   buttonStyleTypes,
   buttonAligns,
   gridRowAlignments,
-  cardStyleTypes
+  cardStyleTypes,
+  progressBarStyleTypes
 };